revert http -> http_session refactor
